requirements.txt
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
behave==1.2.6
behave-django==1.4.0
certifi==2024.2.2
Django==4.2.8
exceptiongroup==1.2.0
factory-boy==3.3.0
Faker==23.2.1
h11==0.14.0
idna==3.6
outcome==1.3.0.post0
parse==1.20.1
parse-type==0.6.2
psycopg2-binary==2.9.9
PySocks==1.7.1
python-dateutil==2.8.2
selenium==4.18.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.1
wsproto==1.2.0
